Rename build_path.js to buildPath.js
I realized this should likely be camelCase.
